+++ utoipa/CHANGELOG.md
@@ -3,6 +3,12 @@
 **`utoipa`** is in direct correlation with **`utoipa-gen`** ([CHANGELOG.md](../utoipa-gen/CHANGELOG.md)). You might want
 to look into changes introduced to **`utoipa-gen`**.
 
+## Unreleased
+
+### Added
+
+* Add `content` to `Parameter` for parameters with complex serialization e.g. JSON-encoded header values
+
 ## 5.4.0 - Jun 16 2025
 
 ### Added